requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile 
#
certifi==2019.3.9         # via requests
chardet==3.0.4            # via requests
click==7.0                # via flask
filelock==3.0.10          # via tox
flask==1.0.2
idna==2.8                 # via requests
itsdangerous==1.1.0       # via flask
jinja2==2.10.1            # via flask
markupsafe==1.1.1         # via jinja2
pluggy==0.11.0            # via tox
py==1.8.0                 # via tox
python-dotenv==0.10.1
requests==2.21.0
six==1.12.0               # via tox, tox-pip-extensions
toml==0.10.0              # via tox
tox-pip-extensions==1.6.0
tox==3.9.0
urllib3==1.24.3           # via requests
virtualenv==16.5.0        # via tox
werkzeug==0.15.2          # via flask
